% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_tracking_targets.R
\name{aggregate_data}
\alias{aggregate_data}
\title{Filters the data set based on the touchstone and date parameters}
\usage{
aggregate_data(dat, params, y_label)
}
\arguments{
\item{dat}{The data set as a data frame}

\item{params}{The parameter set}
}
\value{
A list of three data frames for the projection, actual and future data
}
\description{
Filters the data set based on the touchstone and date parameters
}
