% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{build_meta_data}
\alias{build_meta_data}
\title{Converts a list of parameters into a data frame of metadata}
\usage{
build_meta_data(params)
}
\arguments{
\item{params}{A list of parameters}
}
\value{
A data frame of parameters
}
\description{
Converts a list of parameters into a data frame of metadata
}
