% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_impact_top_countries.R
\name{get_order_of_magnitude}
\alias{get_order_of_magnitude}
\title{Given a number in (0,1e12) calculates the order of magnitude in powers of 1e3}
\usage{
get_order_of_magnitude(x)
}
\arguments{
\item{x}{A numerical value strictly between 0 and 1e12
TODO do we want to catch values outside this range, they should never
occur and if they do it implies there is an error elsewhere in the code}
}
\value{
A list of the wide table and a smaller printable table
}
\description{
Given a number in (0,1e12) calculates the order of magnitude in powers of 1e3
}
