% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_impact_top_countries.R
\name{graph_impact_top_countries}
\alias{graph_impact_top_countries}
\title{Plots the top country impact graph
EXPORTED FUNCTION
Must be called AFTER prep_impact_top_countries}
\usage{
graph_impact_top_countries(dat, params, print_threshold = NA)
}
\arguments{
\item{dat}{A list containing teh filtered data set, and some information
about the units. This is the output from prep_impact_top_countries}

\item{params}{A list (TODO consider using an S3 class) of parameters or the
plot. (TODO does this list have to be the same as the params
for prep_impact_top_countries? If so consider adding params
to dat)}

\item{print_threshold}{A numerical value that determines how when labels are
added to the plot. DEFAULT = NA, which sets the
threshold in the body of the function based on data}
}
\value{
A ggplot2 plot
}
\description{
Plots the top country impact graph
EXPORTED FUNCTION
Must be called AFTER prep_impact_top_countries
}
