% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{graph_num_div}
\alias{graph_num_div}
\title{Create the correct axis label based on the maximum value}
\usage{
graph_num_div(max_result, y_label)
}
\arguments{
\item{max_result}{The maximum value of the y axis}

\item{y_label}{The y axis label to be appended to}
}
\value{
A list contianing the numerical scale (1, 1e3 or 1e6), and the units
in short form as strings  ("", "K", "M") and y label with the long form of
the units appended
}
\description{
Create the correct axis label based on the maximum value
}
