% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_tracking_targets.R
\name{graph_tracking_targets}
\alias{graph_tracking_targets}
\title{Filters and does some processing of the data set before plotting
EXPORTED FUNCTION
Must be called BEFORE graph_impact_top_countries}
\usage{
graph_tracking_targets(dat, params)
}
\arguments{
\item{params}{A list (TODO consider using an S3 class) of parameters for the
plot}

\item{dat_all}{A dataframe (or similar) of country data to be plotted}
}
\value{
A list containing the ggplot2 plot object, a data frame contaiing the
        data used in the plot and some metadata
}
\description{
Filters and does some processing of the data set before plotting
EXPORTED FUNCTION
Must be called BEFORE graph_impact_top_countries
}
