% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_impact_top_countries.R
\name{make_disease_colours}
\alias{make_disease_colours}
\title{Creates a vector of colours for graph_impact_top_countries}
\usage{
make_disease_colours()
}
\value{
A vector of colours
}
\description{
Creates a vector of colours for graph_impact_top_countries
}
