% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_impact_top_countries.R
\name{make_tidy_ylab}
\alias{make_tidy_ylab}
\title{Creates a human readable label for the y axis of the graph_impact_top_countries
plot}
\usage{
make_tidy_ylab(string, unit_label = "")
}
\arguments{
\item{string}{The name of the outcome column in the full data set, usually
"deaths_averted" or "cases_averted"}

\item{unit_label}{A string for the order of magnitude e.g. "thousands",
"millions" or "billions". usualy output from get_order_of_magnitude()}
}
\value{
The y label as a string
}
\description{
Creates a human readable label for the y axis of the graph_impact_top_countries
plot
}
