% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{palette_apply}
\alias{palette_apply}
\alias{palette_vaccine}
\title{Apply a palette to a set of levels}
\usage{
palette_apply(x, pal, missing_error = TRUE, named = TRUE)

palette_vaccine()
}
\arguments{
\item{x}{A character or factor vector}

\item{pal}{A palette to use (named character vector of colours)}

\item{missing_error}{Logical indicating if missing values should throw errors}

\item{named}{Logical indicating if the output should share names with \code{x}}
}
\description{
Apply a palette to a set of levels
}
