% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{params_select}
\alias{params_select}
\title{Filter a data frame (or equivalent) based on a set of parameters}
\usage{
params_select(dat, params)
}
\arguments{
\item{dat}{A data frame to be filtered}

\item{country_set}{A string denoted a set of countries: "gavi73", "gavi69",
"pine" or NULL}

\item{vaccine_type}{A vector of string denoting the type of vaccination type
"campaign" or "routine"}

\item{vaccine}{A vector of string denoting the vaccine "JE", "Measles", "MenA",
"Rubella", "YF", "HepB", "Hib3", "HPV", "MCV2", "PCV3", "Rota"}

\item{country}{A vector of strings of ISO country codes e.g. "BGD"}

\item{continent}{A vector of strings for continents e.g. "Asia"}

\item{region}{A vector of strings for continents e.g. "Southern Asia"}

\item{metric}{A vector of strings for metric "cases_averted", "deaths_averted",
"future_deaths_averted", "future_cases_averted"}

\item{year_lo}{A numeric value the lower bound for the year filtering}

\item{year_hi}{A numeric value the upper bound for the year filtering}

\item{support_type}{A string idincating the support type, usually "gavi" or
"total"}

\item{is_focal}{A boolean, indicating whether or not to use the focal model}

\item{model}{A string indicating which model to use, only used if is_focal is
FALSE}

\item{touchstone}{A vector of string indicating which touchstones to use}
}
\value{
The filtered data frame
}
\description{
Filter a data frame (or equivalent) based on a set of parameters
}
