% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_impact_top_countries.R
\name{prep_impact_top_countries}
\alias{prep_impact_top_countries}
\title{Filters and does some processing of the data set before plotting
EXPORTED FUNCTION
Must be called BEFORE graph_impact_top_countries}
\usage{
prep_impact_top_countries(dat, params)
}
\arguments{
\item{dat}{A dataframe (or similar) of country data for ...}

\item{params}{A list (TODO consider using an S3 class) of parameters or the
plot}
}
\value{
A list containing filtered data, and information about units
        dat, will be a data frame with columns country_name, vaccine_delivery
        and outcome
}
\description{
Filters and does some processing of the data set before plotting
EXPORTED FUNCTION
Must be called BEFORE graph_impact_top_countries
}
