% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_helpers.R
\name{report_figure_with_metadata}
\alias{report_figure_with_metadata}
\title{Make figure with metadata}
\usage{
report_figure_with_metadata(object, filename, meta, ..., data = NULL,
  do_plot = TRUE)
}
\arguments{
\item{object}{A ggplot2 object (must inherit from \code{ggplot})}

\item{filename}{Destination filename to save the plot in.  The
extension must match the device name (e.g.., lowercase)}

\item{meta}{An object of metadata.  Will be saved by running
through \code{yaml::as.yaml}.  Any sort of data should work, but
lists are probably going to be best.}

\item{...}{Additional arguments passed through to the device being
opened. Things like width, height, etc.}

\item{data}{The data used to create the plot.  If not given (which
is the default) the data is pulled from the ggplot object.  This
is probably the right thing to do.}

\item{do_plot}{Logical, indicating if the plot should also be
printed to the null device (the screen, or whatever knitr has
running).}
}
\description{
Helper for producing plots+data+metadata artefacts for orderly.
Only works with ggplot figures, because we use both getting the
data out of the figure (using \code{$data}) and use
\code{print(object)} to produce the plot twice.
}
