% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_helpers.R
\name{report_variable_list}
\alias{report_variable_list}
\title{Create a named list}
\usage{
report_variable_list(...)
}
\arguments{
\item{...}{Things to put in the list}
}
\description{
Helper function for automagically creating a named list the way
that cbind and data.frame do.
}
