% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_tracking_targets.R
\name{set_tracking_title}
\alias{set_tracking_title}
\title{Creates the plot title from the parameter set}
\usage{
set_tracking_title(params)
}
\arguments{
\item{params}{The parameter set}
}
\value{
The plot tile
}
\description{
Creates the plot title from the parameter set
}
