% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_impact_top_countries.R
\name{shorten_name}
\alias{shorten_name}
\title{Simple hacky function that converts longer country names to shorter
alternatives to make the plot look nicer}
\usage{
shorten_name(name)
}
\arguments{
\item{name}{The name of the country as a string}
}
\value{
The shorter (where apllicable) name of the country as a string
}
\description{
Simple hacky function that converts longer country names to shorter
alternatives to make the plot look nicer
}
