% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_impact_top_countries.R
\name{table_impact_top_countries}
\alias{table_impact_top_countries}
\title{Creates a simplified table mirroring the data from the top country impact
graph
EXPORTED}
\usage{
table_impact_top_countries(dat, params)
}
\arguments{
\item{dat}{A list containing the filtered data set, and some information
about the units. This is the output from prep_impact_top_countries}

\item{params}{A list (TODO consider using an S3 class) of parameters or the
plot. (TODO does this list have to be the same as the params
for prep_impact_top_countries? If so consider adding params
to dat)}
}
\value{
A list of the wide table and a smaller printable table
}
\description{
Creates a simplified table mirroring the data from the top country impact
graph
EXPORTED
}
