% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phipps.R
\name{title_case}
\alias{title_case}
\title{Convert a sentence to title case}
\usage{
title_case(str, remove_underscores = FALSE)
}
\arguments{
\item{str}{String (or character vector)}

\item{remove_underscores}{Convert underscores to spaces first}
}
\description{
Convert a sentence to title case
}
