% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_impact_top_countries.R
\name{title_impact_top_countries}
\alias{title_impact_top_countries}
\title{Creates the title for the plot based on the parameters given}
\usage{
title_impact_top_countries(params)
}
\arguments{
\item{params}{A list (TODO consider using an S3 class) of parameters or the
plot.}
}
\value{
The title as an R string
}
\description{
Creates the title for the plot based on the parameters given
}
