% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phipps.R
\name{variable_list}
\alias{variable_list}
\title{Create a named list}
\usage{
variable_list(...)
}
\arguments{
\item{...}{Things to put in the list}
}
\description{
Helper function for automagically creating a named list the way
that cbind and data.frame do.
}
